+++ examples/example_configure_backend.cpp
@@ -10,12 +10,10 @@ int main()
 {
   quill::Config cfg;
 
-#if !defined(QUILL_NO_EXCEPTIONS)
   // Set a custom error handler to handler exceptions - if exceptions are enabled
-  cfg.backend_thread_error_handler = [](std::string const& s)
-  { std::cout << "Hello from error handler. Error: " << s << std::endl; };
-#endif
-  
+  cfg.backend_thread_notification_handler = [](std::string const& s)
+  { std::cout << "Hello from notification handler. Message: " << s << std::endl; };
+
   // Setting to an invalid CPU. When we call quill::start() our error handler will be invoked and an error will be logged
   cfg.backend_thread_cpu_affinity = static_cast<uint16_t>(321312);